# server/openapi/auth.yaml
openapi: 3.0.3
info:
  title: Pet Adoption Auth API
  version: 1.0.0
  description: 用户认证与授权接口
servers:
  - url: http://localhost:8787
paths:
...
  /v1/auth/register:
    post:
      summary: 注册并自动登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                email: { type: string, nullable: true }
                phone: { type: string, nullable: true }
                code: { type: string }
                captcha_id: { type: string }
                captcha_answer: { type: string }
                remember: { type: boolean }
      responses:
        '200':
          description: OK
  /v1/auth/login:
    post:
      summary: 用户名/邮箱/手机号+密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier: { type: string }
                password: { type: string }
                captcha_id: { type: string }
                captcha_answer: { type: string }
                remember: { type: boolean }
      responses:
        '200':
          description: OK
...